+++ scripts/build.sh
@@ -9,7 +9,7 @@ xcrun --sdk iphoneos clang -arch arm64 -isysroot "$sdk" \
   -Werror=implicit-function-declaration -Werror=incompatible-pointer-types -Werror=return-type \
   -install_name '@rpath/QuietTube.dylib' \
   -framework Foundation -framework UIKit \
-  Sources/QTCore.m Sources/QTPreferences.m Sources/QTSettings.m Sources/QTSettingsModel.m Sources/QTFeatures.m Sources/QTLogo.m Sources/QTAdProfile.m Sources/QTMutationTrace.m Sources/QTFeedInsertion.m \
+  Sources/QTCore.m Sources/QTDiagnosticLog.m Sources/QTDiagnosticsBridge.m Sources/QTPreferences.m Sources/QTSettings.m Sources/QTSettingsModel.m Sources/QTFeatures.m Sources/QTLogo.m Sources/QTAdProfile.m Sources/QTMutationTrace.m Sources/QTFeedInsertion.m \
   -o artifacts/QuietTube.dylib
 codesign --force --sign - artifacts/QuietTube.dylib
 file artifacts/QuietTube.dylib